consolidate redundant tests
