+++ platform/flowglad-next/src/server/routers/subscriptionsRouter.integration.test.ts
@@ -45,13 +45,7 @@ import { selectSubscriptionById, updateSubscription } from '@/db/tableMethods/su
 import { selectSubscriptionItems, updateSubscriptionItem } from '@/db/tableMethods/subscriptionItemMethods'
 import { selectBillingPeriodItems } from '@/db/tableMethods/billingPeriodItemMethods'
 import core from '@/utils/core'
-import { selectBillingPeriodById } from '@/db/tableMethods/billingPeriodMethods'
-import { eq } from 'drizzle-orm'
-import { subscriptionItems } from '@/db/schema/subscriptionItems'
-import { attemptToTransitionSubscriptionBillingPeriod } from '@/subscriptions/billingPeriodHelpers'
-import { updateBillingPeriod } from '@/db/tableMethods/billingPeriodMethods'
-import { expireSubscriptionItem } from '@/db/tableMethods/subscriptionItemMethods'
-import { addDays, subDays } from 'date-fns'
+import { subDays } from 'date-fns'
 
 vi.mock('next/headers', () => ({
   headers: vi.fn(() => new Headers()),
@@ -590,464 +584,5 @@ describe('Subscriptions Router - Adjust Endpoint', () => {
         ]),
       })
     }, 10000)
-  })
-})
-
-describe('Billing Period Rollover with Subscription Item Syncing', () => {
-  describe('Subscription Header Sync During Billing Period Transition', () => {
-    test('should sync subscription header when items change during billing period rollover', async () => {
-      await adminTransaction(async ({ transaction }) => {
-        const now = new Date()
-        
-        // 1. Set up subscription with billing period ending in the past to trigger transition
-        const subscriptionWithPastPeriod = await setupSubscription({
-          organizationId: organization.id,
-          customerId: customer.id,
-          priceId: price.id,
-          paymentMethodId: paymentMethod.id,
-          status: SubscriptionStatus.Active,
-          currentBillingPeriodStart: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000), // 30 days ago
-          currentBillingPeriodEnd: new Date(Date.now() - 1000), // 1 second ago - triggers transition
-          renews: true,
-          livemode: true,
-        })
-
-        // 2. Create billing period that has ended
-        const expiredBillingPeriod = await setupBillingPeriod({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          startDate: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          endDate: new Date(Date.now() - 1000), // In the past
-          status: BillingPeriodStatus.Active,
-          livemode: true,
-        })
-
-        // 3. Create old subscription item that should expire
-        const oldItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'Old Basic Plan',
-          quantity: 1,
-          unitPrice: 999, // $9.99
-          addedDate: subDays(now, 20), // Started 20 days ago
-          type: SubscriptionItemType.Static,
-        })
-
-        // 4. Create new subscription item that should become active
-        const newItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: expensivePrice.id,
-          name: 'New Premium Plan',
-          quantity: 1,
-          unitPrice: 4999, // $49.99 - more expensive
-          addedDate: subDays(now, 1), // Started yesterday (now active)
-          type: SubscriptionItemType.Static,
-        })
-
-        // 5. Expire the old item
-        await expireSubscriptionItem(oldItem.id, subDays(now, 1), transaction)
-
-        // 6. Trigger the billing period transition
-        const { result } = await attemptToTransitionSubscriptionBillingPeriod(
-          expiredBillingPeriod,
-          transaction
-        )
-
-        // 7. Verify the subscription header was synced with the new primary item
-        expect(result.subscription.name).toBe('New Premium Plan')
-        expect(result.subscription.priceId).toBe(expensivePrice.id)
-        expect(result.subscription.currentBillingPeriodStart).not.toEqual(expiredBillingPeriod.startDate)
-        expect(result.subscription.currentBillingPeriodEnd).not.toEqual(expiredBillingPeriod.endDate)
-      })
-    })
-
-    test('should handle multiple items becoming active and choose most expensive as primary', async () => {
-      await adminTransaction(async ({ transaction }) => {
-        const now = new Date()
-        
-        // Set up subscription with billing period ending in the past
-        const subscriptionWithPastPeriod = await setupSubscription({
-          organizationId: organization.id,
-          customerId: customer.id,
-          priceId: price.id,
-          paymentMethodId: paymentMethod.id,
-          status: SubscriptionStatus.Active,
-          currentBillingPeriodStart: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          currentBillingPeriodEnd: new Date(Date.now() - 1000),
-          renews: true,
-          livemode: true,
-        })
-
-        const expiredBillingPeriod = await setupBillingPeriod({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          startDate: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          endDate: new Date(Date.now() - 1000),
-          status: BillingPeriodStatus.Active,
-          livemode: true,
-        })
-
-        // Create multiple new items that become active
-        const basicItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'Basic Plan',
-          quantity: 1,
-          unitPrice: 999, // $9.99
-          addedDate: subDays(now, 1),
-          type: SubscriptionItemType.Static,
-        })
-
-        const premiumItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: expensivePrice.id,
-          name: 'Premium Plan',
-          quantity: 1,
-          unitPrice: 4999, // $49.99 - most expensive
-          addedDate: subDays(now, 1),
-          type: SubscriptionItemType.Static,
-        })
-
-        const addonItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'Add-on Feature',
-          quantity: 1,
-          unitPrice: 500, // $5.00
-          addedDate: subDays(now, 1),
-          type: SubscriptionItemType.Static,
-        })
-
-        // Trigger transition
-        const { result } = await attemptToTransitionSubscriptionBillingPeriod(
-          expiredBillingPeriod,
-          transaction
-        )
-
-        // Should choose the most expensive item as primary
-        expect(result.subscription.name).toBe('Premium Plan')
-        expect(result.subscription.priceId).toBe(expensivePrice.id)
-      })
-    })
-
-    test('should handle items with different quantities affecting total price calculations', async () => {
-      await adminTransaction(async ({ transaction }) => {
-        const now = new Date()
-        
-        const subscriptionWithPastPeriod = await setupSubscription({
-          organizationId: organization.id,
-          customerId: customer.id,
-          priceId: price.id,
-          paymentMethodId: paymentMethod.id,
-          status: SubscriptionStatus.Active,
-          currentBillingPeriodStart: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          currentBillingPeriodEnd: new Date(Date.now() - 1000),
-          renews: true,
-          livemode: true,
-        })
-
-        const expiredBillingPeriod = await setupBillingPeriod({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          startDate: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          endDate: new Date(Date.now() - 1000),
-          status: BillingPeriodStatus.Active,
-          livemode: true,
-        })
-
-        // High unit price, low quantity
-        const highUnitPriceItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'High Unit Price',
-          quantity: 1,
-          unitPrice: 5000, // $50 per unit, total = $50
-          addedDate: subDays(now, 1),
-          type: SubscriptionItemType.Static,
-        })
-
-        // Low unit price, high quantity (higher total)
-        const highQuantityItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: expensivePrice.id,
-          name: 'High Quantity',
-          quantity: 10,
-          unitPrice: 1000, // $10 per unit, total = $100 (MORE expensive)
-          addedDate: subDays(now, 1),
-          type: SubscriptionItemType.Static,
-        })
-
-        const { result } = await attemptToTransitionSubscriptionBillingPeriod(
-          expiredBillingPeriod,
-          transaction
-        )
-
-        // Should choose high quantity item (higher total)
-        expect(result.subscription.name).toBe('High Quantity')
-        expect(result.subscription.priceId).toBe(expensivePrice.id)
-      })
-    })
-  })
-
-  describe('Future Adjustments During Billing Period Rollover', () => {
-    test('should sync subscription header when future adjustments become active during rollover', async () => {
-      await adminTransaction(async ({ transaction }) => {
-        const now = new Date()
-        
-        // Set up subscription with billing period ending in the past
-        const subscriptionWithPastPeriod = await setupSubscription({
-          organizationId: organization.id,
-          customerId: customer.id,
-          priceId: price.id,
-          paymentMethodId: paymentMethod.id,
-          status: SubscriptionStatus.Active,
-          currentBillingPeriodStart: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          currentBillingPeriodEnd: new Date(Date.now() - 1000),
-          renews: true,
-          livemode: true,
-        })
-
-        const expiredBillingPeriod = await setupBillingPeriod({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          startDate: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          endDate: new Date(Date.now() - 1000),
-          status: BillingPeriodStatus.Active,
-          livemode: true,
-        })
-
-        // Create items that were scheduled for "AtEndOfCurrentBillingPeriod" timing
-        // These should become active exactly when the billing period rolls over
-        const futureItem1 = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'Future Basic Plan',
-          quantity: 1,
-          unitPrice: 1999, // $19.99
-          addedDate: subDays(now, 1), // Started yesterday (now active)
-          type: SubscriptionItemType.Static,
-        })
-
-        const futureItem2 = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: expensivePrice.id,
-          name: 'Future Premium Plan',
-          quantity: 1,
-          unitPrice: 6999, // $69.99 - most expensive
-          addedDate: subDays(now, 1), // Started yesterday (now active)
-          type: SubscriptionItemType.Static,
-        })
-
-        // Trigger the billing period transition
-        const { result } = await attemptToTransitionSubscriptionBillingPeriod(
-          expiredBillingPeriod,
-          transaction
-        )
-
-        // Should sync with the most expensive future item
-        expect(result.subscription.name).toBe('Future Premium Plan')
-        expect(result.subscription.priceId).toBe(expensivePrice.id)
-      })
-    })
-
-    test('should handle mixed immediate and future adjustments during rollover', async () => {
-      await adminTransaction(async ({ transaction }) => {
-        const now = new Date()
-        
-        const subscriptionWithPastPeriod = await setupSubscription({
-          organizationId: organization.id,
-          customerId: customer.id,
-          priceId: price.id,
-          paymentMethodId: paymentMethod.id,
-          status: SubscriptionStatus.Active,
-          currentBillingPeriodStart: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          currentBillingPeriodEnd: new Date(Date.now() - 1000),
-          renews: true,
-          livemode: true,
-        })
-
-        const expiredBillingPeriod = await setupBillingPeriod({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          startDate: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          endDate: new Date(Date.now() - 1000),
-          status: BillingPeriodStatus.Active,
-          livemode: true,
-        })
-
-        // Items that were already active (immediate adjustments)
-        const immediateItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'Immediate Plan',
-          quantity: 1,
-          unitPrice: 2999, // $29.99
-          addedDate: subDays(now, 10), // Started 10 days ago
-          type: SubscriptionItemType.Static,
-        })
-
-        // Items that become active during rollover (future adjustments)
-        const futureItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: expensivePrice.id,
-          name: 'Future Premium Plan',
-          quantity: 1,
-          unitPrice: 7999, // $79.99 - most expensive overall
-          addedDate: subDays(now, 1), // Started yesterday
-          type: SubscriptionItemType.Static,
-        })
-
-        const { result } = await attemptToTransitionSubscriptionBillingPeriod(
-          expiredBillingPeriod,
-          transaction
-        )
-
-        // Should choose the most expensive item overall
-        expect(result.subscription.name).toBe('Future Premium Plan')
-        expect(result.subscription.priceId).toBe(expensivePrice.id)
-      })
-    })
-  })
-
-  describe('Complex Rollover Scenarios', () => {
-    test('should handle items becoming active and inactive simultaneously during rollover', async () => {
-      await adminTransaction(async ({ transaction }) => {
-        const now = new Date()
-        
-        const subscriptionWithPastPeriod = await setupSubscription({
-          organizationId: organization.id,
-          customerId: customer.id,
-          priceId: price.id,
-          paymentMethodId: paymentMethod.id,
-          status: SubscriptionStatus.Active,
-          currentBillingPeriodStart: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          currentBillingPeriodEnd: new Date(Date.now() - 1000),
-          renews: true,
-          livemode: true,
-        })
-
-        const expiredBillingPeriod = await setupBillingPeriod({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          startDate: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          endDate: new Date(Date.now() - 1000),
-          status: BillingPeriodStatus.Active,
-          livemode: true,
-        })
-
-        // Old items that should expire
-        const oldBasicItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'Old Basic Plan',
-          quantity: 1,
-          unitPrice: 999,
-          addedDate: subDays(now, 20),
-          type: SubscriptionItemType.Static,
-        })
-
-        const oldPremiumItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: expensivePrice.id,
-          name: 'Old Premium Plan',
-          quantity: 1,
-          unitPrice: 4999,
-          addedDate: subDays(now, 20),
-          type: SubscriptionItemType.Static,
-        })
-
-        // New items that become active
-        const newBasicItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'New Basic Plan',
-          quantity: 1,
-          unitPrice: 1999, // $19.99
-          addedDate: subDays(now, 1),
-          type: SubscriptionItemType.Static,
-        })
-
-        const newPremiumItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: expensivePrice.id,
-          name: 'New Premium Plan',
-          quantity: 1,
-          unitPrice: 6999, // $69.99 - most expensive overall
-          addedDate: subDays(now, 1),
-          type: SubscriptionItemType.Static,
-        })
-
-        const newAddonItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'New Add-on',
-          quantity: 1,
-          unitPrice: 500, // $5.00
-          addedDate: subDays(now, 1),
-          type: SubscriptionItemType.Static,
-        })
-
-        // Expire old items
-        await expireSubscriptionItem(oldBasicItem.id, subDays(now, 1), transaction)
-        await expireSubscriptionItem(oldPremiumItem.id, subDays(now, 1), transaction)
-
-        const { result } = await attemptToTransitionSubscriptionBillingPeriod(
-          expiredBillingPeriod,
-          transaction
-        )
-
-        // Should use the most expensive new item
-        expect(result.subscription.name).toBe('New Premium Plan')
-        expect(result.subscription.priceId).toBe(expensivePrice.id)
-      })
-    })
-
-    test('should maintain subscription state when no items are active after rollover', async () => {
-      await adminTransaction(async ({ transaction }) => {
-        const now = new Date()
-        
-        const subscriptionWithPastPeriod = await setupSubscription({
-          organizationId: organization.id,
-          customerId: customer.id,
-          priceId: price.id,
-          paymentMethodId: paymentMethod.id,
-          status: SubscriptionStatus.Active,
-          currentBillingPeriodStart: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          currentBillingPeriodEnd: new Date(Date.now() - 1000),
-          renews: true,
-          livemode: true,
-        })
-
-        const expiredBillingPeriod = await setupBillingPeriod({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          startDate: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
-          endDate: new Date(Date.now() - 1000),
-          status: BillingPeriodStatus.Active,
-          livemode: true,
-        })
-
-        // Create an item that was active but expires before rollover
-        const expiredItem = await setupSubscriptionItem({
-          subscriptionId: subscriptionWithPastPeriod.id,
-          priceId: price.id,
-          name: 'Expired Plan',
-          quantity: 1,
-          unitPrice: 2999,
-          addedDate: subDays(now, 10),
-          type: SubscriptionItemType.Static,
-        })
-
-        // Expire the item before rollover
-        await expireSubscriptionItem(expiredItem.id, subDays(now, 2), transaction)
-
-        const { result } = await attemptToTransitionSubscriptionBillingPeriod(
-          expiredBillingPeriod,
-          transaction
-        )
-
-        // Should maintain the current subscription state when no items are active
-        // The subscription record itself should remain unchanged
-        expect(result.subscription.id).toBe(subscriptionWithPastPeriod.id)
-        expect(result.subscription.customerId).toBe(subscriptionWithPastPeriod.customerId)
-        expect(result.subscription.organizationId).toBe(subscriptionWithPastPeriod.organizationId)
-        // The subscription should have new billing period dates
-        expect(result.subscription.currentBillingPeriodStart).not.toEqual(expiredBillingPeriod.startDate)
-        expect(result.subscription.currentBillingPeriodEnd).not.toEqual(expiredBillingPeriod.endDate)
-      })
-    })
   })
 })
