+++ platform/flowglad-next/src/subscriptions/adjustSubscription.test.ts
@@ -2,6 +2,7 @@ import { describe, it, expect, beforeEach } from 'vitest'
 import {
   adjustSubscription,
   calculateSplitInBillingPeriodBasedOnAdjustmentDate,
+  syncSubscriptionWithActiveItems,
 } from '@/subscriptions/adjustSubscription'
 import {
   BillingPeriodStatus,
@@ -31,15 +32,17 @@ import {
   selectCurrentBillingPeriodForSubscription,
   updateBillingPeriod,
 } from '@/db/tableMethods/billingPeriodMethods'
-import { updateSubscription } from '@/db/tableMethods/subscriptionMethods'
 import { selectBillingPeriodItems } from '@/db/tableMethods/billingPeriodItemMethods'
+import { selectBillingRuns } from '@/db/tableMethods/billingRunMethods'
+import { updateSubscription } from '@/db/tableMethods/subscriptionMethods'
+import { updateSubscriptionItem, expireSubscriptionItem } from '@/db/tableMethods/subscriptionItemMethods'
+import { addDays, subDays } from 'date-fns'
 import { SubscriptionItem } from '@/db/schema/subscriptionItems'
 import { Customer } from '@/db/schema/customers'
 import { PaymentMethod } from '@/db/schema/paymentMethods'
 import { BillingPeriod } from '@/db/schema/billingPeriods'
 import { BillingRun } from '@/db/schema/billingRuns'
 import { Subscription } from '@/db/schema/subscriptions'
-import { selectBillingRuns } from '@/db/tableMethods/billingRunMethods'
 
 describe('adjustSubscription Integration Tests', async () => {
   const { organization, price } = await setupOrg()
@@ -1254,6 +1257,376 @@ describe('adjustSubscription Integration Tests', async () => {
         // Even with proration, subscription record should be updated
         expect(result.subscription.name).toBe('Updated Plan - Pro')
         expect(result.subscription.priceId).toBe(price.id)
+      })
+    })
+  })
+
+  /* ==========================================================================
+     syncSubscriptionWithActiveItems Tests
+  ========================================================================== */
+  describe('syncSubscriptionWithActiveItems', () => {
+    it('should sync subscription with currently active items', async () => {
+      await adminTransaction(async ({ transaction }) => {
+        const now = new Date()
+        const futureDate = addDays(now, 1) // Tomorrow
+        
+        // Setup: Create current active item
+        const currentItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Current Plan',
+          quantity: 1,
+          unitPrice: 999, // $9.99
+          addedDate: subDays(now, 10), // Started 10 days ago
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Expire the current item in the future
+        await expireSubscriptionItem(currentItem.id, futureDate, transaction)
+        
+        // Setup: Create future item that will become active tomorrow
+        const futureItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'New Premium Plan',
+          quantity: 1,
+          unitPrice: 4999, // $49.99
+          addedDate: futureDate, // Starts tomorrow
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Test: Sync should use current item (future item not active yet)
+        const synced = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        expect(synced.name).toBe('Current Plan')
+        expect(synced.priceId).toBe(currentItem.priceId)
+      })
+    })
+
+    it('should handle multiple items becoming active and choose the most expensive as primary', async () => {
+      await adminTransaction(async ({ transaction }) => {
+        const now = new Date()
+        const pastDate = subDays(now, 1) // Yesterday
+        
+        // Setup: Create multiple items that are all currently active
+        await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Basic Feature',
+          quantity: 1,
+          unitPrice: 500, // $5.00
+          addedDate: pastDate,
+          type: SubscriptionItemType.Static,
+        })
+        
+        const premiumItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Premium Feature',
+          quantity: 1,
+          unitPrice: 3000, // $30.00 - Most expensive
+          addedDate: pastDate,
+          type: SubscriptionItemType.Static,
+        })
+        
+        await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Standard Feature',
+          quantity: 1,
+          unitPrice: 1500, // $15.00
+          addedDate: pastDate,
+          type: SubscriptionItemType.Static,
+        })
+        
+        // All items are active now - should choose the most expensive (Premium Feature)
+        const synced = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        
+        expect(synced.name).toBe('Premium Feature')
+        expect(synced.priceId).toBe(premiumItem.priceId)
+      })
+    })
+
+    it('should handle subscription becoming active but not primary (lower price than existing)', async () => {
+      await adminTransaction(async ({ transaction }) => {
+        const now = new Date()
+        const futureDate = addDays(now, 1)
+        
+        // Setup: Create expensive current item
+        const expensiveItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Enterprise Plan',
+          quantity: 1,
+          unitPrice: 9999, // $99.99 - Most expensive
+          addedDate: subDays(now, 10),
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Setup: Create cheaper item that is also active
+        await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Add-on Feature',
+          quantity: 1,
+          unitPrice: 999, // $9.99 - Cheaper
+          addedDate: subDays(now, 1), // Already active since yesterday
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Both are active now - should still use Enterprise Plan as primary
+        const synced = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        
+        expect(synced.name).toBe('Enterprise Plan')
+        expect(synced.priceId).toBe(expensiveItem.priceId)
+      })
+    })
+
+    it('should update primary when current primary item gets cancelled', async () => {
+      await adminTransaction(async ({ transaction }) => {
+        const now = new Date()
+        
+        // Setup: Create multiple active items
+        const primaryItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Premium Plan',
+          quantity: 1,
+          unitPrice: 4999, // $49.99 - Initially most expensive
+          addedDate: subDays(now, 10),
+          type: SubscriptionItemType.Static,
+        })
+        
+        const secondaryItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Standard Plan',
+          quantity: 1,
+          unitPrice: 2999, // $29.99 - Second most expensive
+          addedDate: subDays(now, 5),
+          type: SubscriptionItemType.Static,
+        })
+        
+        await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Basic Plan',
+          quantity: 1,
+          unitPrice: 999, // $9.99 - Cheapest
+          addedDate: subDays(now, 3),
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Initial sync - should use Premium Plan
+        const syncedBefore = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        expect(syncedBefore.name).toBe('Premium Plan')
+        
+        // Cancel the primary item - set as already expired
+        await updateSubscriptionItem({
+          id: primaryItem.id,
+          expiredAt: subDays(now, 1), // Already expired yesterday
+          type: SubscriptionItemType.Static,
+        }, transaction)
+        
+        // Sync after cancellation - should switch to Standard Plan
+        const syncedAfter = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        expect(syncedAfter.name).toBe('Standard Plan')
+        expect(syncedAfter.priceId).toBe(secondaryItem.priceId)
+      })
+    })
+
+    it('should handle multiple items becoming active and inactive simultaneously', async () => {
+      await adminTransaction(async ({ transaction }) => {
+        const now = new Date()
+        
+        // Setup: Currently active items (old items)
+        await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Old Basic',
+          quantity: 1,
+          unitPrice: 999,
+          addedDate: subDays(now, 10),
+          type: SubscriptionItemType.Static,
+        })
+        
+        const oldPremiumItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Old Premium',
+          quantity: 1,
+          unitPrice: 4999,
+          addedDate: subDays(now, 10),
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Setup: New items that are also active now (simulating post-rollover state)
+        await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'New Basic',
+          quantity: 1,
+          unitPrice: 1999, // $19.99
+          addedDate: subDays(now, 1), // Started yesterday
+          type: SubscriptionItemType.Static,
+        })
+        
+        const newPremiumItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'New Premium',
+          quantity: 1,
+          unitPrice: 6999, // $69.99 - Most expensive overall
+          addedDate: subDays(now, 1), // Started yesterday
+          type: SubscriptionItemType.Static,
+        })
+        
+        await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'New Add-on',
+          quantity: 1,
+          unitPrice: 500, // $5.00
+          addedDate: subDays(now, 1), // Started yesterday
+          type: SubscriptionItemType.Static,
+        })
+        
+        // With all items active - should use New Premium (most expensive)
+        const synced = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        expect(synced.name).toBe('New Premium')
+        expect(synced.priceId).toBe(newPremiumItem.priceId)
+      })
+    })
+
+    it('should maintain subscription state when all items expire with no replacements', async () => {
+      await adminTransaction(async ({ transaction }) => {
+        const now = new Date()
+        
+        // Setup: Create an active item first
+        const activeItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Active Plan',
+          quantity: 1,
+          unitPrice: 2999,
+          addedDate: subDays(now, 10),
+          type: SubscriptionItemType.Static,
+        })
+        
+        // First, sync while the item is active to set the subscription name
+        const syncedActive = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        expect(syncedActive.name).toBe('Active Plan')
+        
+        // Now expire the item
+        await updateSubscriptionItem({
+          id: activeItem.id,
+          expiredAt: subDays(now, 1), // Already expired
+          type: SubscriptionItemType.Static,
+        }, transaction)
+        
+        // Sync after expiration with no active items
+        const syncedAfterExpiry = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        
+        // Should maintain the last known state (Active Plan)
+        expect(syncedAfterExpiry.name).toBe('Active Plan')
+        expect(syncedAfterExpiry.priceId).toBe(price.id)
+        expect(syncedAfterExpiry.id).toBe(subscription.id)
+      })
+    })
+
+    it('should handle quantity changes affecting total price calculations', async () => {
+      await adminTransaction(async ({ transaction }) => {
+        const now = new Date()
+        
+        // Setup: Item with high unit price but low quantity
+        const highUnitPriceItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'High Unit Price',
+          quantity: 1,
+          unitPrice: 5000, // $50 per unit, total = $50
+          addedDate: subDays(now, 5),
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Setup: Item with low unit price but high quantity
+        const highQuantityItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'High Quantity',
+          quantity: 10,
+          unitPrice: 1000, // $10 per unit, total = $100 (MORE expensive)
+          addedDate: subDays(now, 5),
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Sync - should choose high quantity item (higher total)
+        const synced = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        
+        expect(synced.name).toBe('High Quantity')
+        expect(synced.priceId).toBe(highQuantityItem.priceId)
+      })
+    })
+
+    it('should use addedDate as tiebreaker when items have same total price', async () => {
+      await adminTransaction(async ({ transaction }) => {
+        const now = new Date()
+        
+        // Setup: Two items with same total price
+        await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Older Item',
+          quantity: 1,
+          unitPrice: 3000, // $30.00
+          addedDate: subDays(now, 10), // Older
+          type: SubscriptionItemType.Static,
+        })
+        
+        const newerItem = await setupSubscriptionItem({
+          subscriptionId: subscription.id,
+          priceId: price.id,
+          name: 'Newer Item',
+          quantity: 1,
+          unitPrice: 3000, // Same price: $30.00
+          addedDate: subDays(now, 5), // Newer - should win
+          type: SubscriptionItemType.Static,
+        })
+        
+        // Sync - should choose newer item as tiebreaker
+        const synced = await syncSubscriptionWithActiveItems(
+          subscription.id,
+          transaction
+        )
+        
+        expect(synced.name).toBe('Newer Item')
+        expect(synced.priceId).toBe(newerItem.priceId)
       })
     })
   })
